requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Classful==0.14.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
paho-mqtt==1.5.0
python-dotenv==0.14.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.3.3
